pyproject.toml
[tool.poetry]
name = "superdebug"
version = "0.1.16"
description = "Convenient debugging for machine learning projects"
authors = ["Azure-Vision <hewanrong2001@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
mypyc-ipython = "^0.0.2"
Pillow = "^5.0.0"
torch = "^1.0.0"
numpy = "^1.0.0"
torchvision = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
